makeTodo, signupPage, and uploadPage restructured
