+++ src/pages/SignUpPage/_sign-up.scss
@@ -1,0 +1,77 @@
+.signup {
+  @include flex(column nowrap);
+  width: 100vw;
+  height: 100vh;
+  color: var(--black, $black);
+  background: var(--white, $white);
+  form {
+    @include border();
+    display: flex;
+    background: var(--white, $white);
+    flex-flow: column nowrap;
+    align-items: center;
+    justify-content: space-between;
+    padding: 1rem;
+    width: 80%;
+    height: 80%;
+    max-height: 600px;
+    border-radius: 5px;
+    label {
+      height: 7%;
+      width: 100%;
+      font-size: 1.1rem;
+      input {
+        width: 100%;
+        height: 100%;
+        border: 1px solid red;
+        border-radius: 5px;
+        font-size: 1.4rem;
+      }
+    }
+    .pass-label {
+      div {
+        width: 100%;
+        height: 100%;
+        display: flex;
+        flex-flow: row nowrap;
+        justify-content: space-between;
+        align-items: center;
+        .sign-pass {
+          width: 80%;
+        }
+        .pass-visibility {
+          @include flex(row nowrap);
+          width: 5%;
+          height: 50%;
+          background: transparent;
+          border: none;
+          cursor: pointer;
+        }
+        .icon {
+          @include flex(row nowrap);
+          border: none;
+          width: 30px;
+          height: 30px;
+          border-radius: 50%;
+        }
+      }
+    }
+    input[type="submit"] {
+      font-size: 1.2rem;
+      border: none;
+      height: 7%;
+      border-radius: 5px;
+      width: 70%;
+      color: var(--white, $white);
+      transition: transform 1s;
+      background: $green;
+      cursor: pointer;
+    }
+    input[type="submit"]:hover {
+      transform: scale(1.1);
+    }
+    input[type="submit"]:active {
+      transform: scale(0.5);
+    }
+  }
+}
